Update project to latest tag sacerping-jboss-8.0.0

+++ preingest-ejb/src/main/java/it/eng/sacerasi/sisma/ejb/InvioSismaEjb.java
@@ -14,9 +14,9 @@
  * You should have received a copy of the GNU Affero General Public License along with this program.
  * If not, see <https://www.gnu.org/licenses/>.
  */
-
 package it.eng.sacerasi.sisma.ejb;
 
+import it.eng.parer.objectstorage.dto.BackendStorage;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -57,6 +57,7 @@ import it.eng.sacerasi.entity.PigObject;
 import it.eng.sacerasi.entity.PigSisma;
 import it.eng.sacerasi.entity.PigSisma.TiStato;
 import it.eng.sacerasi.entity.PigSismaDocumenti;
+import it.eng.sacerasi.entity.PigSismaDocumentiStorage;
 import it.eng.sacerasi.entity.PigSismaFaseProgetto;
 import it.eng.sacerasi.entity.PigSismaFinanziamento;
 import it.eng.sacerasi.entity.PigSismaProgettiAg;
@@ -76,10 +77,12 @@ import it.eng.sacerasi.sisma.xml.invioSisma.ProgettiSisma.Ente;
 import it.eng.sacerasi.sisma.xml.invioSisma.ProgettiSisma.TipiDocumento;
 import it.eng.sacerasi.sisma.xml.invioSisma.ProgettiSisma.TipiDocumento.TipoDocumento;
 import it.eng.sacerasi.sisma.xml.invioSisma.ProgettiSisma.TipoDocumentoPrincipale;
+import it.eng.sacerasi.versamento.ejb.VersamentoOggettoEjb;
 import it.eng.sacerasi.viewEntity.PigVSismaChecks;
 import it.eng.sacerasi.web.ejb.AmministrazioneEjb;
 import it.eng.sacerasi.web.helper.AmministrazioneHelper;
 import it.eng.sacerasi.web.helper.ConfigurationHelper;
+import it.eng.sacerasi.web.util.Utils;
 import it.eng.sacerasi.ws.ejb.XmlContextCache;
 import it.eng.sacerasi.ws.invioOggettoAsincrono.ejb.InvioOggettoAsincronoEjb;
 import it.eng.sacerasi.ws.notificaTrasferimento.dto.FileDepositatoType;
@@ -87,6 +90,7 @@ import it.eng.sacerasi.ws.notificaTrasferimento.dto.ListaFileDepositatoType;
 import it.eng.sacerasi.ws.notificaTrasferimento.ejb.NotificaTrasferimentoEjb;
 import it.eng.sacerasi.ws.response.InvioOggettoAsincronoRisposta;
 import it.eng.sacerasi.ws.response.NotificaTrasferimentoRisposta;
+import java.util.Optional;
 import software.amazon.awssdk.core.ResponseInputStream;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.services.s3.model.GetObjectResponse;
@@ -129,6 +133,8 @@ public class InvioSismaEjb {
     private AmministrazioneEjb amministrazioneEjb;
     @EJB
     private SalvataggioBackendHelper salvataggioBackendHelper;
+    @EJB
+    private VersamentoOggettoEjb versamentoOggettoEjb;
 
     public void invioSisma() {
         log.info("{} --- Chiamata JOB per invio sisma", InvioSismaEjb.class.getSimpleName());
@@ -258,42 +264,46 @@ public class InvioSismaEjb {
                         // Recupero i file dall'object storage
                         List<PigSismaDocumenti> documentiDaInviare = invioSismaHelper
                                 .getDocumentiDaInviare(sismaDaInviare.getIdSisma());
-                        //
-                        ObjectStorageBackend config = salvataggioBackendHelper.getObjectStorageConfiguration("SISMA",
-                                configurationHelper.getValoreParamApplicByApplic(Constants.BUCKET_VERIFICA_SISMA));
 
-                        if (salvataggioBackendHelper.isActive()) {
-                            for (PigSismaDocumenti sismaDocumenti : documentiDaInviare) {
-                                String nmFileOs = sismaDocumenti.getNmFileOs();
-                                String nmFileOrig = sismaDocumenti.getNmFileOrig();
-                                // Chiamata di tipo HEAD (non contiene il body in quanto ho bisogno di una sola
-                                // informazione
-                                // stile ack)
-                                boolean doesObjectExist = salvataggioBackendHelper.doesObjectExist(config, nmFileOs);
-                                if (doesObjectExist) {
-                                    File tempFile = File.createTempFile(nmFileOrig, "", new File(dirCompletaFtp));
-                                    try (FileOutputStream fosTemp = new FileOutputStream(tempFile);) {
-                                        ResponseInputStream<GetObjectResponse> object = salvataggioBackendHelper
-                                                .getObject(config, nmFileOs);
-                                        // Creo il file in una cartella temporanea
-                                        IOUtils.copy(object, fosTemp);
-                                        String nomeFileLowerCase = sismaDocumenti.getPigSismaValDoc()
-                                                .getNmTipoDocumento().replace(" ", "_").toLowerCase() + ".zip";
-                                        // Aggiungo il file scaricato dall'OS al file zip
-                                        addToZipFile(tempFile, zos, nomeFileLowerCase);
-                                    } finally {
-                                        if (tempFile != null) {
-                                            FileUtils.deleteQuietly(tempFile);
-                                        }
+                        for (PigSismaDocumenti sismaDocumenti : documentiDaInviare) {
+                            // MEV 34843
+                            PigSismaDocumentiStorage pigSismaDocumentiStorage = sismaDocumenti
+                                    .getPigSismaDocumentiStorage();
+                            BackendStorage backend = salvataggioBackendHelper
+                                    .getBackend(pigSismaDocumentiStorage.getIdDecBackend());
+                            ObjectStorageBackend config = salvataggioBackendHelper
+                                    .getObjectStorageConfigurationForSisma(backend.getBackendName(),
+                                            pigSismaDocumentiStorage.getNmBucket());
+
+                            String nmFileOs = pigSismaDocumentiStorage.getCdKeyFile();
+                            String nmFileOrig = sismaDocumenti.getNmFileOrig();
+                            // Chiamata di tipo HEAD (non contiene il body in quanto ho bisogno di una sola
+                            // informazione
+                            // stile ack)
+                            boolean doesObjectExist = salvataggioBackendHelper.doesObjectExist(config, nmFileOs);
+                            if (doesObjectExist) {
+                                File tempFile = File.createTempFile(nmFileOrig, "", new File(dirCompletaFtp));
+                                try (FileOutputStream fosTemp = new FileOutputStream(tempFile);) {
+                                    ResponseInputStream<GetObjectResponse> object = salvataggioBackendHelper
+                                            .getObject(config, nmFileOs);
+                                    // Creo il file in una cartella temporanea
+                                    IOUtils.copy(object, fosTemp);
+                                    String nomeFileLowerCase = sismaDocumenti.getPigSismaValDoc().getNmTipoDocumento()
+                                            .replace(" ", "_").toLowerCase() + ".zip";
+                                    // Aggiungo il file scaricato dall'OS al file zip
+                                    addToZipFile(tempFile, zos, nomeFileLowerCase);
+                                } finally {
+                                    if (tempFile != null) {
+                                        FileUtils.deleteQuietly(tempFile);
                                     }
-                                } else {
-                                    log.error("{} --- ERRORE creazione ZIP invio sisma",
-                                            InvioSismaEjb.class.getSimpleName());
-                                    PigErrore errore = messaggiHelper.retrievePigErrore("PING-ERRSISMA04");
-                                    String dsErrore = StringUtils.replace(errore.getDsErrore(), "{0}",
-                                            sismaDocumenti.getNmFileOs());
-                                    throw new InvioSismaException(idSismaDaInviare, errore.getCdErrore(), dsErrore);
                                 }
+                            } else {
+                                log.error("{} --- ERRORE creazione ZIP invio sisma",
+                                        InvioSismaEjb.class.getSimpleName());
+                                PigErrore errore = messaggiHelper.retrievePigErrore("PING-ERRSISMA04");
+                                String dsErrore = StringUtils.replace(errore.getDsErrore(), "{0}",
+                                        sismaDocumenti.getPigSismaDocumentiStorage().getCdKeyFile());
+                                throw new InvioSismaException(idSismaDaInviare, errore.getCdErrore(), dsErrore);
                             }
                         }
                     } catch (Exception ex) {
@@ -346,13 +356,34 @@ public class InvioSismaEjb {
                         PigErrore errore = messaggiHelper.retrievePigErrore(ERR_01);
                         throw new InvioSismaException(idSismaDaInviare, errore.getCdErrore(), errore.getDsErrore());
                     }
-                    String zipSisma = sismaDaInviare.getCdKey();
+
+                    // MEV 34843
+                    PigTipoObject pigTipoObject = amministrazioneHelper.getPigTipoObjectByName(
+                            InvioSismaHelper.NOME_TIPO_OGGETTO_DA_TRASFORMARE, new BigDecimal(vers.getIdVers()));
+                    BigDecimal idAmbiente = BigDecimal.valueOf(vers.getPigAmbienteVer().getIdAmbienteVers());
+                    BigDecimal idVers = BigDecimal.valueOf(vers.getIdVers());
+                    BigDecimal idTipoObject = BigDecimal.valueOf(pigTipoObject.getIdTipoObject());
+                    BackendStorage backendVersamento = salvataggioBackendHelper.getBackendForVersamento(idAmbiente,
+                            idVers, idTipoObject);
+                    String nmFileOs = null;
+
                     try {
-                        // Effettua lo scarico del nuovo file da trasformare nella cartella ftp di SacerPing dell'utente
-                        // relativo nella cartella INPUT_FOLDER
-                        NfsUtils.createEmptyDir(dirCompletaFtp + zipSisma);
-                        Files.move(Paths.get(dirCompletaFtp + nomeFilePacchetto + ".TEMP"),
-                                Paths.get(String.join(zipSisma, dirCompletaFtp, "/", ".zip")));
+                        if (backendVersamento.isFile()) {
+                            // Effettua lo scarico del nuovo file da trasformare nella cartella ftp di SacerPing
+                            // dell'utente
+                            // relativo nella cartella INPUT_FOLDER
+                            NfsUtils.createEmptyDir(dirCompletaFtp + cdKeyObject);
+                            Files.move(Paths.get(dirCompletaFtp + nomeFilePacchetto + ".TEMP"),
+                                    Paths.get(String.join(cdKeyObject, dirCompletaFtp, "/", ".zip")));
+                        } else if (backendVersamento.isObjectStorage()) {
+                            // MEV 34843
+                            ObjectStorageBackend config = salvataggioBackendHelper
+                                    .getObjectStorageConfigurationForVersamento(backendVersamento.getBackendName());
+                            nmFileOs = versamentoOggettoEjb.computeOsFileKey(idVers, cdKeyObject,
+                                    VersamentoOggettoEjb.OS_KEY_POSTFIX.PIGOBJECT.name());
+                            salvataggioBackendHelper.putS3Object(config, nmFileOs, fileTemporaneoGenerale,
+                                    Optional.empty());
+                        }
                     } catch (IOException ex) {
                         log.error("{}{}", InvioSismaEjb.class.getSimpleName(), ERR_INVIO_SISMA);
                         PigErrore errore = messaggiHelper.retrievePigErrore("PING-ERRSISMA15");
@@ -382,6 +413,21 @@ public class InvioSismaEjb {
                             fileDepositatoType.setNmTipoFile(NOME_TIPO_FILE);
                             fileDepositatoType.setNmNomeFile(nomeFilePacchetto);
                             listaFileDepositatoType.getFileDepositato().add(fileDepositatoType);
+
+                            // MEV 34843
+                            fileDepositatoType.setIdBackend(backendVersamento.getBackendId());
+
+                            if (backendVersamento.isObjectStorage()) {
+                                ObjectStorageBackend config = salvataggioBackendHelper
+                                        .getObjectStorageConfigurationForVersamento(backendVersamento.getBackendName());
+                                String tenantOs = configurationHelper.getValoreParamApplicByApplic(
+                                        it.eng.sacerasi.common.Constants.TENANT_OBJECT_STORAGE);
+
+                                fileDepositatoType.setNmOsBucket(config.getBucket());
+                                fileDepositatoType.setNmOsTenant(tenantOs);
+                                fileDepositatoType.setNmNomeFileOs(nmFileOs);
+                            }
+
                             risposta = notificaTrasferimentoEjb.notificaAvvenutoTrasferimentoFileInNewTx(nmAmbienteVers,
                                     nmVers, cdKeyObject, listaFileDepositatoType);
                         } catch (Exception e) {
@@ -397,13 +443,23 @@ public class InvioSismaEjb {
                     // Il sistema effettua il caricamento del file ZipSisma nel Bucket
                     // BUCKET_SISMA_TRASFORMATI
                     try {
-                        ObjectStorageBackend config = salvataggioBackendHelper.getObjectStorageConfiguration("SISMA",
-                                configurationHelper.getValoreParamApplicByApplic(Constants.BUCKET_VERIFICA_SISMA));
-                        if (salvataggioBackendHelper.isActive()) {
-                            salvataggioBackendHelper.putS3Object(config, sismaDaInviare.getCdKeyOs() + ".zip", new File(
-                                    rootFtp + "/" + dsPathInputFtp + "/" + zipSisma + "/" + zipSisma + ".zip"));
+                        BackendStorage backend = salvataggioBackendHelper.getBackendForSisma();
+
+                        if (backend.isObjectStorage()) {
+                            ObjectStorageBackend config = salvataggioBackendHelper
+                                    .getObjectStorageConfigurationForSismaTrasformati(backend.getBackendName());
+                            if (fileTemporaneoGenerale.exists()) { // il backend è su object storage e
+                                                                   // fileTemporaneoGenerale esiste ancora.
+                                salvataggioBackendHelper.putS3Object(config, sismaDaInviare.getCdKeyOs() + ".zip",
+                                        fileTemporaneoGenerale, Optional.empty());
+                            } else { // il backend è su disco e fileTemporaneoGenerale è stato spostato nella sua
+                                     // posizione definitiva.
+                                salvataggioBackendHelper.putS3Object(config, sismaDaInviare.getCdKeyOs() + ".zip",
+                                        Paths.get(String.join(cdKeyObject, dirCompletaFtp, "/", ".zip")).toFile(),
+                                        Optional.empty());
+                            }
                         }
-                    } catch (SdkClientException e) {
+                    } catch (Exception e) {
                         log.error("{}{}{}", InvioSismaEjb.class.getSimpleName(), " --- ERRORE invio sisma: ",
                                 e.getMessage());
                         PigErrore errore = messaggiHelper.retrievePigErroreNewTx("PING-ERRSISMA19");
